requirement.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astropy==5.2.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
cachetools==4.2.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.12
decorator==5.2.1
docker-pycreds==0.4.0
dominate==2.6.0
dpcpp-cpp-rt==2025.0.4
easydict==1.13
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
face-alignment==1.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.0.0
graphviz==0.8.4
grpcio==1.68.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
intel-cmplr-lib-rt==2025.0.4
intel-cmplr-lib-ur==2025.0.4
intel-cmplr-lic-rt==2025.0.4
intel-opencl-rt==2025.0.4
intel-openmp==2025.0.4
intel-sycl-rt==2025.0.4
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.11.9
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mkl==2025.0.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
mxnet==1.9.1
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.3
numpy==1.23.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnx2torch==1.5.15
onnxruntime==1.19.2
onnxsim==0.4.36
opencv-python==3.4.9.33
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.0.0
prettytable==3.11.0
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyDeprecate==0.3.1
pyerfa==2.0.0.3
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-image==0.16.2
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sphinx_glpi_theme==0.6
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchmetrics==1.5.2
torchvision==0.19.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==3.9.20
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.1
umf==0.9.1
urllib3==2.2.3
wandb==0.19.1
wcwidth==0.2.13
Werkzeug==3.0.6
yacs==0.1.8
yarl==1.15.2
zipp==3.20.2
